+++ packages/contracts-cdn/package.json
@@ -7,9 +7,8 @@
 		"lint:fix": "bunx @biomejs/biome lint --write src/",
 		"format": "dprint check",
 		"format:fix": "dprint fmt --verbose",
-		"token:generate": "bun scripts/generate-token-list.ts",
 		"cache:create": "bun scripts/create-cache.ts",
-		"cache:test": "bun scripts/test-cache.ts",
+		"compile:benchmark": "bun scripts/compile-benchmark.ts",
 		"deploy": "wrangler deploy",
 		"dev": "wrangler dev",
 		"start": "wrangler dev",
