+++ packages/contracts-cdn/scripts/compile-benchmark.ts
