+++ packages/contracts-cdn/wrangler.toml
@@ -1,6 +1,6 @@
 #:schema node_modules/wrangler/config-schema.json
 name                = "luminadex-contracts-cdn"
-main                = "src/test-cache.ts"
+main                = "src/index.ts"
 compatibility_date  = "2024-11-27"
 compatibility_flags = ["nodejs_compat"]
